+++ CHANGELOG.md
@@ -73,6 +73,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix issue where Select All could stop working in Tracker view
 - Fix issue where multiple context menus could be opened at once
 - Fix issue where colors could appear differently in palette editor to how they appear in game
+- Fix issue where changing one color in a palette could also slightly change the other colors
 
 ## [4.2.2] - 2026-02-24